Fixed Compiler error/warning

+++ src/MyContactListener.cpp
@@ -8,11 +8,11 @@ void MyContactListener::BeginContact(b2Contact* contact)
 {
     //check if fixture A was the foot sensor
     void* fixtureUserData = contact->GetFixtureA()->GetUserData();
-    if ( (int)fixtureUserData == 3 )
+    if ( (intptr_t)fixtureUserData == 3 )
         numFootContacts++;
     //check if fixture B was the foot sensor
     fixtureUserData = contact->GetFixtureB()->GetUserData();
-    if ( (int)fixtureUserData == 3 )
+    if ( (intptr_t)fixtureUserData == 3 )
         numFootContacts++;
 }
 
@@ -20,10 +20,10 @@ void MyContactListener::EndContact(b2Contact* contact)
 {
     //check if fixture A was the foot sensor
     void* fixtureUserData = contact->GetFixtureA()->GetUserData();
-    if ( (int)fixtureUserData == 3 )
+    if ( (intptr_t)fixtureUserData == 3 )
         numFootContacts--;
     //check if fixture B was the foot sensor
     fixtureUserData = contact->GetFixtureB()->GetUserData();
-    if ( (int)fixtureUserData == 3 )
+    if ( (intptr_t)fixtureUserData == 3 )
         numFootContacts--;
 }
